페이지를 만드는 로직 component에 추가하는 방식으로 변경
